-karma +webpack
Also, move some code around
