+++ .gitignore
@@ -1,3 +1,2 @@
 *.log
 node_modules
-integration/.*
