+++ index.js
@@ -1,4 +1,4 @@
-import check from './lib/check';
-import register from './lib/register';
+import check from './src/check';
+import register from './src/register';
 
 export { check, register };
