+++ src/check/index.js
