+++ src/check/spec.js
